[refactoring] color theme custom hook
깔끔한 커스텀 훅
