+++ src/components/Mobile/Entry/EntryFooter/OtherPostMobile/OtherPostMobile.tsx
@@ -1,4 +1,3 @@
-import { colorThemeAtom } from "@/utils/atoms";
 import {
   Box,
   Center,
@@ -6,14 +5,11 @@ import {
   VStack,
   useColorModeValue,
 } from "@chakra-ui/react";
-import { useRecoilValue } from "recoil";
-import { useEffect } from "react";
-import { returnColors } from "@/utils/utilFn";
-import { useState } from "react";
 import LeftBoxMobile from "./LeftBoxMobile";
 import RightBoxMobile from "./RightBoxMobile";
 import OtherPostProfileMobile from "./OtherPostProfileMobile";
 import { IOtherPostProps } from "@/components/Entry/EntryFooter/OtherPost/OtherPost";
+import { useColorTheme } from "@/hooks/useColorTheme";
 
 export interface IBoxProps {
   relativeColor: string;
@@ -23,18 +19,8 @@ export interface IBoxProps {
 }
 
 export default function OtherPostMobile({ next, prev }: IOtherPostProps) {
-  const colorTheme = useRecoilValue(colorThemeAtom);
+  const { relativeColor } = useColorTheme();
   const mdBgColor = useColorModeValue("#F9F8FA", "rgba(45,55,72,1)");
-  const [lightColor, setLightColor] = useState("");
-  const [darkColor, setDarkColor] = useState("");
-  const relativeColor = useColorModeValue(lightColor, darkColor);
-
-  useEffect(() => {
-    const [lc, dc, hbc] = returnColors(colorTheme);
-    setLightColor(lc);
-    setDarkColor(dc);
-  }, [colorTheme]);
-
   return (
     <>
       <Center w="full" h={"auto"} pt={44}>
