+++ src/components/Mobile/GuestBook/GuestBookMainPageMobile.tsx
@@ -1,18 +1,8 @@
-import { colorThemeAtom } from "@/utils/atoms";
-import { returnColors } from "@/utils/utilFn";
-import { Box, Center, Heading, Image, Text, VStack } from "@chakra-ui/react";
-import { useEffect, useState } from "react";
-import { useRecoilValue } from "recoil";
+import { Box, VStack } from "@chakra-ui/react";
 import GBInputMobile from "./GBInputMobile";
 import GBCommentsMobile from "./GBCommentsMobile";
 
 export default function GuestBookMainPageMobile() {
-  const colorTheme = useRecoilValue(colorThemeAtom);
-  const [lightColor, setLightColor] = useState("");
-  useEffect(() => {
-    const [lc, dc, hbc] = returnColors(colorTheme);
-    setLightColor(lc);
-  }, [colorTheme]);
   return (
     <>
       <VStack
